adicionando mensagens de erro nos controller

+++ src/controllers/OrderController.js
@@ -66,7 +66,7 @@ class OrderController {
   static async show(req, res) {
     const { id } = req.params;
 
-    const orders = await Order.findByPk(id, {
+    const order = await Order.findByPk(id, {
       include: [
         {
           association: 'products',
@@ -87,6 +87,10 @@ class OrderController {
       attributes: { exclude: ['tab_id'] },
     });
 
+    if (order == null) {
+      return res.status(400).json({ error: 'Order not found.' });
+    }
+
     return res.status(201).json(orders);
   }
 
@@ -102,6 +106,12 @@ class OrderController {
         },
       }
     );
+
+    const order = await Order.findByPk(id);
+
+    if (order == null) {
+      return res.status(400).json({ error: 'Order not found.' });
+    }
 
     const productsOrder = await ProductsOrdered.findAll({
       where: {
@@ -137,6 +147,12 @@ class OrderController {
 
   static async delete(req, res) {
     const { id } = req.params;
+
+    const order = await Order.findByPk(id);
+
+    if (order == null) {
+      return res.status(400).json({ error: 'Order not found.' });
+    }
 
     await Order.destroy({
       where: {
